business-insight-chatbot/streamlit_sales_insights_agent/requirements.txt
streamlit~=1.38.0
pandas~=2.2.3
mlflow
openai>=1.0.0
databricks-sql-connector